pyproject.toml
[build-system]
build-backend = 'setuptools.build_meta'
requires = [
    'setuptools >= 61.2',
    'wheel'
]

[project]
name = 'Tilda'
version = '1.25'
description = 'A python package for the LaserSpHERe experiment control and data analysis.'
readme = 'README.md'
requires-python = '>=3.8'
license = {file = 'LICENSE'}
keywords = ['physics', 'laser spectroscopy', 'data acquisition', 'DAQ', 'data analysis']
authors = [
    {name='LaserSpHERe group', email='lasersphere_exp@ikp.tu-darmstadt.de'}
]
maintainers = [
  {name = 'Laura Renth', email = 'lrenth@ikp.tu-darmstadt.de'},
  {name = 'Patrick Mueller', email = 'pamueller@ikp.tu-darmstadt.de'}
]
classifiers = [
    'Programming Language :: Python :: 3',
    'License :: OSI Approved :: MIT License',
    'Operating System :: Microsoft :: Windows'
]
dependencies = [
    'cffi >= 1.14.6',
    'comtypes >= 1.1.10',
    'cryptography >= 3.4.7',
    'cycler >= 0.10.0',
    'h5py >= 3.3.0',
    'kiwisolver >= 1.3.1',
    'lxml >= 4.6.3',
    'matplotlib >= 3.4.3',
    'mysql-connector-python >= 8.0.32',
    'numpy >= 1.21.2',
    'Pillow >= 8.3.1',
    'protobuf >= 3.17.3',
    'pycparser >= 2.20',
    'pyparsing >= 2.4.7',
    'PyQt5 >= 5.15.4',
    'PyQt5-Qt5 >= 5.15.2',
    'PyQt5-sip >= 12.9.0',
    'pyqtgraph >= 0.12.2',
    'pyserial >= 3.5',
    'python-dateutil >= 2.8.2',
    'pytz >= 2021.1',
    'PyVISA >= 1.11.3',
    'PyVISA-py >= 0.5.2',
    'pywin32 >= 301',
    'PyYAML >= 5.4.1',
    'scipy >= 1.7.1',
    'six >= 1.16.0',
    'typing-extensions >= 3.10.0.0',
    'requests >= 2.28.2',
    'influxdb >= 5.3.1'
]

[project.urls]
homepage = 'https://www.ikp.tu-darmstadt.de/forschung_kernphysik/gruppen_kernphysik/experiment/ag_w_noertershaeuser/index.en.jsp'
repository = 'http://gitlab.ikp.physik.tu-darmstadt.de/Lasersphere/Tilda'
bug-tracker = 'http://gitlab.ikp.physik.tu-darmstadt.de/Lasersphere/Tilda/issues'

[tool.setuptools.package-data]
'*' = ['*.dll', '*.lvbitx', '*.wav', '*.jpg', '*.ico']
